requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-bootstrap4==0.0.8
django-braces==1.13.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.10.2
gunicorn==19.9.0
psycopg2==2.8.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.3
